Change extension functions to prefix commands (hidden) Fixes #9
