--- src/data/channels.json.example
@@ -1,5 +1,0 @@
-{
-	"guild_id_here" : {
-		"chan_id_here" : "chan_owner_id_here"
-	}
-}
