+++ src/data/guilds.json
@@ -1,0 +1,10 @@
+{
+    "952930870858240031": {
+        "cat": 960992907962232862,
+        "chan": 960992908901752862
+    },
+    "956966239736049725": {
+        "cat": 960992911581929542,
+        "chan": 960992912634699806
+    }
+}
