--- src/data/guilds.json.example
@@ -1,6 +1,0 @@
-{
-	"guild_id_here" : {
-		"cat" : cat_id_here,
-		"chan" : chan_id_here
-	}
-}
