+++ src/extensions/slash_commands.py
@@ -8,36 +8,6 @@ class SlashCommands(commands.Cog):
     def __init__(self, bot):
         self.bot:commands.Bot = bot
 
-    @commands.slash_command(description="Load an extension")
-    @commands.is_owner()
-    async def load(self, inter: disnake.ApplicationCommandInteraction, path: str):
-        """ Load an extension """
-        try:
-            self.bot.load_extension(path)
-            await inter.response.send_message(content=f"{path} was loaded.", ephemeral=True)
-        except (commands.NoEntryPointError, commands.ExtensionNotFound):
-            await inter.response.send_message(content=f"Could not find an extension with name {path}.", ephemeral=True)
-        except commands.ExtensionAlreadyLoaded:
-            try:
-                self.bot.reload_extension(path)
-                await inter.response.send_message(content=f"{path} was reloaded.", ephemeral=True)
-            except commands.ExtensionFailed:
-                await inter.response.send_message(content="Something went wrong.", ephemeral=True)
-        except commands.ExtensionFailed:
-            await inter.response.send_message(content="Something went wrong.", ephemeral=True)
-
-    @commands.slash_command(description="Unload an extension")
-    @commands.is_owner()
-    async def unload(self, inter: disnake.ApplicationCommandInteraction, path: str):
-        """ Unload an extension """
-        try:
-            self.bot.unload_extension(path)
-            await inter.response.send_message(content=f"{path} was unloaded.", ephemeral=True)
-        except commands.ExtensionNotFound:
-            await inter.response.send_message(contant=f"Could not find an extension with name {path}.", ephemeral=True)
-        except commands.ExtensionNotLoaded:
-            await inter.response.send_message(content=f"{path} was not loaded.", ephemeral=True)
-
     @commands.slash_command(description="Clear n messages")
     @commands.has_permissions(administrator=True)
     async def wipe(self, inter: disnake.ApplicationCommandInteraction, n: int):
